refactor: 마이히스토리 파일 위치 변경, 로딩/에러 처리
