+++ src/components/PageComponents/tasks/TaskError.tsx
@@ -1,6 +1,6 @@
-export default function TaskListSkeleton() {
+export default function TaskError() {
   return (
-    <div className="mx-auto mt-96 text-center text-text-md text-text-default sm:mt-56">
+    <div className="pt-[17rem] text-center text-text-md font-medium text-text-default sm:pt-[11rem] md:pt-[14rem]">
       <p>할 일을 불러오는데 에러가 발생했습니다.</p>
       <p>잠시 후 다시 접속해주세요.</p>
     </div>
